pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "BrieFlow"
version = "1.5.0"
description = "Package for processing data from optical poooled screens."
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
  "snakemake==8.30.0",
  "snakemake-executor-plugin-slurm==1.4.0",
  "sphinx-rtd-theme==3.0.2",
  "ipykernel==6.29.5",
  "pandas==2.3.0",
  "numpy>=1.26",
  "pyyaml>=6.0",
  "nd2==0.10.3",
  "scikit-image==0.24.0",
  "joblib==1.5.1",
  "cellpose==3.1.0",
  "csbdeep==0.8.1",
  "microfilm==0.3.0",
  "seaborn==0.13.2",
  "tifffile==2025.6.11",
  "mahotas==1.4.18",
  "cp-measure==0.1.12",
  "python-igraph==0.11.9",
  "leidenalg==0.10.2",
  "phate==1.0.11",
  "statsmodels==0.14.4",
  "dill==0.4.0",
  "xgboost==3.0.4",
  "ruff==0.12.1",
  "streamlit==1.45.1",
  "plotly==6.1.2",
  "pytest==8.4.0",
  "matplotlib==3.10.3",
  "adjusttext==1.3.0",
  "torch==2.7.1",
  "spotiflow==0.5.7",
  "stardist==0.9.1",
  "dexp==2023.4.10.686",
  "levenshtein==0.27.1",
  "ipywidgets==8.1.7",
  "jupyterlab-widgets==3.0.15",
  "widgetsnbextension==4.0.14",
  "lightgbm==4.6.0",
  "scikit-learn==1.8.0",
  "ome-zarr>=0.8.0",
  "zarr>=2.16.0",
  "dask[array]",
  "fsspec",
]


[tool.setuptools]
package-dir = {"" = "workflow"}
packages = ["lib"]

[tool.ruff]

[tool.ruff.lint]
# Configure pydocstyle settings under lint
select = ["D"]
pydocstyle = { convention = "google" }

[tool.ruff.lint.per-file-ignores]
# Ignore docstring check for script and notebook files
"workflow/scripts/*.py" = ["D100", "D103"]
"*.ipynb" = ["D100", "D103"]
# Ignore all checks for external files
"workflow/lib/external/*.py" = ["ALL"]
# Ignore all checks for docs
"docs/source/*.py" = ["ALL"]
# Ignore all checks for tests
"tests/*.py" = ["ALL"]
# Visualization code
"visualization/*.py" = ["ALL"]

[tool.pytest.ini_options]
markers = [
  "unit: unit tests",
  "integration: integration tests",
]
